fix: q4 coalesced gemv scale tensor is f32, not f16 (#33)
mt_gemv_q4_coalesced, _relu2, and _accum declared their scale input as
`d_f32: Tensor<f16>`, a leftover from the kernel-family reorg. Every
signal says the intended dtype is f32: the parameter is named `d_f32`,
the in-file benches bind it as `DType::F32` (so they were feeding f32
bytes into an f16-typed slot), and every sibling reads f32 scales
(mt_gemv_q8_coalesced and its variants, plus mt_gemv_q4_coalesced_2row
and mt_gemv_q4_vec). The f16 typing made the kernel read f32 scale
bytes as packed halfs, producing garbage output.

There is no q4-coalesced correctness test, which is how the drift went
unnoticed; gemv_q8_correctness and kernel_registry_consistency pass
unchanged, and the benches now type-match their f32 buffers. The body
already does `load(d_f32[...]).cast::<f32>()`, so reading f32 is a
no-op cast.

+++ crates/metaltile-std/src/kernels/gemm/gemv_quantized.rs
@@ -238,7 +238,7 @@ pub fn mt_gemv_q8_coalesced_accum<T>(
 #[kernel]
 pub fn mt_gemv_q4_coalesced<T>(
     qs: Tensor<u32>,
-    d_f32: Tensor<f16>,
+    d_f32: Tensor<f32>,
     x: Tensor<T>,
     mut out: Tensor<T>,
     #[constexpr] k_in: u32,
@@ -432,7 +432,7 @@ pub fn mt_gemv_q4_coalesced_2row<T>(
 #[kernel]
 pub fn mt_gemv_q4_coalesced_relu2<T>(
     qs: Tensor<u32>,
-    d_f32: Tensor<f16>,
+    d_f32: Tensor<f32>,
     x: Tensor<T>,
     mut out: Tensor<T>,
     #[constexpr] k_in: u32,
@@ -482,7 +482,7 @@ pub fn mt_gemv_q4_coalesced_relu2<T>(
 #[allow(clippy::too_many_arguments)]
 pub fn mt_gemv_q4_coalesced_accum<T>(
     qs: Tensor<u32>,
-    d_f32: Tensor<f16>,
+    d_f32: Tensor<f32>,
     x: Tensor<T>,
     mut acc: Tensor<T>,
     scale: Tensor<f32>,
